Move exschemapath proto.
