+++ protomap/proto_test.go
@@ -22,7 +22,7 @@ import (
 	"google.golang.org/protobuf/proto"
 
 	wpb "github.com/openconfig/ygot/proto/ywrapper"
-	epb "github.com/openconfig/ygot/ygot/testdata/exschemapath"
+	epb "github.com/openconfig/ygot/protomap/testdata/exschemapath"
 )
 
 func TestPathsFromProto(t *testing.T) {
