+++ protomap/testdata/exschemapath/gen.go
